requirements/base.txt
djangorestframework==3.15.2
Markdown==3.7
django-environ==0.11.2
psycopg==3.2.2
pandas==2.2.3